fix: seperation of concern
