+++ apps/feedahelp/src/components/ui/features/deshiMenus/Cta/index.ts
@@ -1,0 +1,1 @@
+export * from "./Cta"
