+++ apps/feedahelp/src/components/ui/features/deshiMenus/ExtraInformation/index.ts
@@ -1,0 +1,1 @@
+export * from "./ExtraInformation"
