+++ apps/feedahelp/src/components/ui/features/deshiMenus/MainContent/index.ts
@@ -1,0 +1,1 @@
+export * from "./MainContent"
